requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
ollama==0.1.7
python-dotenv==1.0.0
pandas==2.1.3
requests==2.31.0

# LangChain for agent implementation
langchain==0.1.0
langchain-community==0.0.13
langchain-ollama==0.0.1
sqlalchemy==2.0.25

pytest==7.4.3
httpx==0.25.2